requirements.txt
Django==2.1
django-js-reverse==0.8.2
djangorestframework==3.8.2
